requirements.txt
ag-ui-protocol==0.1.10
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.11.0
argcomplete==3.6.3
attrs==25.4.0
authlib==1.6.5
beartype==0.22.5
beautifulsoup4==4.8.2
boto3==1.40.74
botocore==1.40.74
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.0
cohere==5.20.0
colorama==0.4.6
compressed-rtf==1.0.7
cryptography==46.0.3
cyclopts==4.2.4
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
docx2txt==0.9
ebcdic==1.1.1
email-validator==2.3.0
eval-type-backport==0.3.0
exceptiongroup==1.3.0
executing==2.2.1
extract-msg==0.29.0
fastapi==0.121.2
fastavro==1.12.1
fastmcp==2.13.0.2
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.10.0
genai-prices==0.0.39
google-auth==2.43.0
google-genai==1.50.1
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio-client==1.13.3
griffe==1.15.0
groovy==0.1.2
groq==0.34.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.1.4
idna==3.11
imapclient==2.1.0
importlib-metadata==8.7.0
invoke==2.2.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jeepney==0.9.0
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
logfire==4.14.2
logfire-api==4.14.2
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.21.1
mdurl==0.1.2
mistralai==1.9.11
more-itertools==10.8.0
nexus-rpc==1.1.0
numpy==2.3.4
olefile==0.47
openai==2.8.0
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pdfminer-six==20191110
pillow==11.3.0
pip==24.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
protobuf==6.33.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-ai==1.18.0
pydantic-ai-slim==1.18.0
pydantic-core==2.33.2
pydantic-evals==1.18.0
pydantic-graph==1.18.0
pydantic-settings==2.12.0
pydub==0.25.1
pygments==2.19.2
pyjwt==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.5
s3transfer==0.14.0
safehttpx==0.1.7
secretstorage==3.4.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
speechrecognition==3.8.1
sse-starlette==3.0.3
starlette==0.49.3
temporalio==1.19.0
tenacity==9.1.2
textract==1.6.5
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
types-protobuf==6.32.1.20251105
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
word2number==1.1
wrapt==1.17.3
xlrd==1.2.0
xlsxwriter==3.2.9
zipp==3.23.0
